pyproject.toml
[project]
name = "lsp-code-fix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "genlm-control",
    "pyright",
    "dspy>=0.0.0",  # update version if you want the latest
    "transformers",
    "torch"
]
